separated static and dynamic configs due to collision
